issue-2: create auth service controller and repository
