+++ controllers/user.controller.js
@@ -64,40 +64,11 @@ const deleteUserByUsername = async (req, res) => {
 
 };
 
-const userRegistration = async(req,res)=>{
-
-  try {
-    const userRegistration = await userService.userRegistration(
-      req.body
-    );
-    res.status(200).json(userRegistration);
-  } catch (err) {
-    console.error(err);
-    res.send({ status: 500, message: "Internal Server Error" });
-  }
-  
-}
-
-const userLogIn = async(req,res)=>{
-
-  try {
-    const userLogIn = await userService.userLogIn(
-      req.body
-    );
-    res.status(200).json(userLogIn);
-  } catch (err) {
-    console.error(err);
-    res.send({ status: 500, message: "Internal Server Error" });
-  }
-  
-}
 
 module.exports = {
   getAllUser,
   createUser,
   getUserByUsername,
   updateUserPasswordByUsername,
-  deleteUserByUsername,
-  userRegistration,
-  userLogIn
+  deleteUserByUsername
 };
