+++ routes/auth.route.js
@@ -1,10 +1,10 @@
-const userController = require("../controllers/user.controller");
+const authController = require("../controllers/auth.controller");
 
 const express = require("express");
 const router = express.Router();
 
 
 router.route("/register").post(userController.userRegistration);
-router.route("/login").post(userController.userLogIn);
+router.route("/login").post(authController.userLogIn);
 
 module.exports = router;
