+++ utils/auth.utils.js
@@ -1,0 +1,6 @@
+exports.comparePassword = async (inputPassword,userPassword) => {
+    const comparePasswordResult = await bcrypt.compare(inputPassword,userPassword);
+    return comparePasswordResult;
+  };
+
+
